(PC-16625) refactor(profile): reorganise profile folders
