+++ src/features/cheatcodes/pages/NavigationProfile/NavigationProfile.tsx
@@ -4,13 +4,10 @@ import styled from 'styled-components/native'
 
 import { LinkToComponent } from 'features/cheatcodes/components/LinkToComponent'
 import { useGoBack } from 'features/navigation/useGoBack'
-import {
-  domains_credit_v1,
-  domains_credit_v2,
-} from 'features/profile/components/headers/fixtures/domainsCredit'
-import { CreditCeilingsModal } from 'features/profile/components/headers/Modals/CreditCeilingsModal'
-import { ExhaustedCreditModal } from 'features/profile/components/headers/Modals/ExhaustedCreditModal'
-import { ExpiredCreditModal } from 'features/profile/components/headers/Modals/ExpiredCreditModal'
+import { CreditCeilingsModal } from 'features/profile/components/Modals/CreditCeilingsModal'
+import { ExhaustedCreditModal } from 'features/profile/components/Modals/ExhaustedCreditModal'
+import { ExpiredCreditModal } from 'features/profile/components/Modals/ExpiredCreditModal'
+import { domains_credit_v1, domains_credit_v2 } from 'features/profile/fixtures/domainsCredit'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { ModalHeader } from 'ui/components/modals/ModalHeader'
 import { useModal } from 'ui/components/modals/useModal'
