+++ src/features/profile/components/Header/CreditHeader/CreditHeader.tsx
@@ -3,9 +3,9 @@ import React from 'react'
 import styled from 'styled-components/native'
 
 import { DomainsCredit } from 'api/gen/api'
+import { BeneficiaryCeilings } from 'features/profile/components/BeneficiaryCeilings/BeneficiaryCeilings'
 import { CreditExplanation } from 'features/profile/components/CreditExplanation/CreditExplanation'
 import { CreditInfo } from 'features/profile/components/CreditInfo/CreditInfo'
-import { BeneficiaryCeilings } from 'features/profile/components/headers/BeneficiaryCeilings/BeneficiaryCeilings'
 import { useIsUserUnderageBeneficiary } from 'features/profile/utils'
 import { formatToSlashedFrenchDate } from 'libs/dates'
 import { PageHeader } from 'ui/components/headers/PageHeader'
