--- src/features/profile/components/__mocks__/NonBeneficiaryHeader.tsx
@@ -1,3 +1,0 @@
-export function NonBeneficiaryHeader() {
-  return 'NonBeneficiaryHeader'
-}
