+++ src/features/profile/pages/NotificationSettings.tsx
@@ -8,7 +8,7 @@ import styled from 'styled-components/native'
 import { NotificationSubscriptions, UserProfileResponse } from 'api/gen'
 import { useAuthContext } from 'features/auth/AuthContext'
 import { useUserProfileInfo } from 'features/profile/api'
-import { SectionWithSwitch } from 'features/profile/components/SectionWithSwitch'
+import { SectionWithSwitch } from 'features/profile/components/SectionWithSwitch/SectionWithSwitch'
 import { useAppStateChange } from 'libs/appState'
 import { analytics } from 'libs/firebase/analytics'
 import { PushNotificationsModal } from 'libs/notifications/components/PushNotificationsModal'
