+++ src/features/profile/tests/api.test.ts
@@ -3,7 +3,7 @@ import { rest } from 'msw'
 import { UserProfileResponse } from 'api/gen'
 import { useAuthContext } from 'features/auth/AuthContext'
 import { useUserProfileInfo } from 'features/profile/api'
-import { domains_credit_v1 } from 'features/profile/components/headers/fixtures/domainsCredit'
+import { domains_credit_v1 } from 'features/profile/fixtures/domainsCredit'
 import { env } from 'libs/environment'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { server } from 'tests/server'
